requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.1
gunicorn==20.0.4
channels==3.0.3
django-redis==4.12.1
django-crispy-forms==1.11.2
django-ckeditor==6.0.0
Pillow==8.1.2
feedparser==6.0.2
beautifulsoup4==4.9.3
requests
selenium
itsdangerous
django-allauth
six